pyproject.toml
[tool.poetry]
name = "kchromik-hackathon-open-meteo-challenge"
version = "0.0.1"
description = ""
authors = ["JangasCodingplace <janisgoesser92@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 100

[tool.mypy]
ignore_missing_imports = true

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic-settings = "^2.3.4"
pytz = "^2024.1"
geopy = "^2.4.1"
timezonefinder = "^6.5.2"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
sqlmodel = "^0.0.19"
django = "^5.0.7"
djangorestframework = "^3.15.2"
jinja2 = "^3.1.4"
django-cors-headers = "^4.4.0"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.1"
isort = "^5.13.2"
black = "^24.4.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
